requirement.txt
numpy
pandas
sklearn.preprocessing
seaborn 
matplotlib.pyplot